Defer initialization of Crispin connection pool so that it can be retried.
CrispinConnectionPool.__init__ calls CrispinConnectionPool._set_account_info(),
which does auth and related stuff. We were calling this in the FolderSyncEngine
constructor, which meant that the folder sync would just silently die if some
sort of error occurred. Not good.

In general we really shouldn't be doing this sort of work in any class
constructors at all.