Only IDLE on the inbox folder for Gmail.
Summary:
This fixes a more than slightly mortifying regression in which we were idling
on ALL folders that we sync for Gmail. The problem with this is that it was
consuming all available crispin connections. Because the sync on the inbox
folder goes and gets a new all_mail_crispin_client to download threads, this
was causing the polling on the inbox folder to potentially block for an
unbounded amount of time. No bueno.

Test Plan:
Verify in sync logs that only the inbox folder sync idles, and that
it is not blocked.

Reviewers: spang

Reviewed By: spang

Subscribers: dlitz

Differential Revision: https://review.inboxapp.com/D510