[SCHEMA] Properly reconcile Inbox-created messages in IMAP sync.
Summary:
If we sync a message with an X-INBOX-ID header and already have it in that
namespace's data store, don't create a new message. Just update the existing
thread instead.

Fixes a bug in which a message sent via the Inbox API causes duplicate threads
to appear once you sync the sent message from the backend.

Note: Requires migration, but can be run online. Takes <15 minutes.

Test Plan:
Check that syncing a sent message updates the existing thread,
instead of creating an old one.

Reviewers: spang, kav-ya

Reviewed By: kav-ya

Differential Revision: https://review.inboxapp.com/D528