Create canonical tag objects at Namespace creation time.
Summary:
We assume that tags like 'sent' and 'drafts'are always available, for example
in the drafts/sending API functionality, but we previously only ensured
that they existed when we started a sync. This only caused problems.

Instead, just create the tag objects when a new Namespace is initialized.

Test Plan: Run unit tests; auth and sync an account.

Reviewers: kav-ya

Reviewed By: kav-ya

Differential Revision: https://review.inboxapp.com/D659