Fix T569.
Summary:
resync_uids() was correctly asking for an additinal paramter... self.
However, being a closure, it takes self directly from the environment,
thus we don't need to declare it.

Test Plan:
Just sync a bunch of accounts, we should not encounter the TypeError reported
in https://review.inboxapp.com/T569 anymore. We should see a
NotImplementedError, which should be gracefully by the retry wrapper.

Reviewers: dlitz, khamidou, emfree

Reviewed By: emfree

Subscribers: spang

Differential Revision: https://review.inboxapp.com/D677