Calendar bug fixes
Summary:
This diff fixes a bunch of bugs with the Gmail calendars.

The most egregious one is a weird issue with the mapping we used to assign participants to an event. I changed it to use one of SQLAlchemy's default mapping.

There's also a couple fixes in the way we handle changes in events -- previously we'd barf if an event field was updated on the server. I modified the merging strategy to "barf if setting an field to None". I think this should be okay most of the time and eventually consistent.

Finally, I changed the gmail update code to export more of our db fields.

Also, I think I'm going to release code like this, in daily batches. This should make thing easier to review.

Test Plan: Tested manually.

Reviewers: kav-ya, emfree

Reviewed By: kav-ya, emfree

Subscribers: spang

Projects: #sync-engine

Differential Revision: https://review.inboxapp.com/D693