Add a system test to check for expected broken accounts
Summary:
This is the system test for D765 and finishes up the sync engine side
of T495 - checking for All Mail folder and failing gracefully if it's
absent.

This test specifically adds another check in `test_auth` based on a
new list of live, bad credentials in accounts.py. It's purposefully
general, so other known bad credentials could be added if we want to
test for more error cases.

There are two changes required in Jenkins to make this run:

* Add another string parameter broken_accounts to identify
  broken test accounts.
* In the build script, after the line:
  `echo "credentials = $test_accounts" > tests/system/accounts.py`
  add:
  `echo "broken_credentials = $broken_accounts" >> tests/system/accounts.py`

It looks like I can make the changes myself but I'd appreciate input
from folks more familiar with the jenkins setup.

Test Plan: If you want to test locally, ping me for some broken credentials you can use.

Reviewers: emfree

Reviewed By: emfree

Subscribers: spang

Differential Revision: https://review.inboxapp.com/D773