Remove deleted UIDs after adding new ones in IMAP sync.
Summary:
Say we sync a draft update back to the IMAP backend, so delete the UID for the
old draft and add a new one. In normal mailsync, if we first sync the delete
and then the add, the odds are high that the actual draft object gets deleted
as being orphaned, and that we then create a new different draft object once we
sync the add. This would cause a baffling 404 error if you went to send the
draft or update it again. (T621)

As aesthetically unappealing as this solution is, it drastically lessens the
likelihood of that scenario in an expedient way.

Test Plan:
Repeatedly update draft and check that sync runs correctly without
deleting it.

Reviewers: spang

Reviewed By: spang

Differential Revision: https://review.inboxapp.com/D794