Improve logging for Contacts and Events sync.
Summary:
This revision improves logging for Contacts and Events sync in case a
`ConnectionError` occurs. Up to now, we're relying on the heartbeat status
which correctly signals the greenlet is in the `poll error` state. However,
there is nothing in our logs that says what's going on.

NOTE: this revison also removes the `except PermissionError` turning it into an
uncaught error since we now start contacts and events sync if and only if we
have the permissions to access them.

Fix 703.

Test Plan:
Sync a Gmail account with permissions to access mails, contacts, and events.
Inhibit access to either contacts and events (e.g., change `/etc/hosts`) and
make sure `ConnectionError`s get logged.

Reviewers: khamidou, emfree

Reviewed By: emfree

Differential Revision: https://review.inboxapp.com/D959