Make Accounts unique
Summary: This is the sister diff to D1106. Mostly, I added a unique constraint to the `_canonicalized_address` field of the `Account` table. Please note that to do this I had to redefine the field in the `Account` table, otherwise SQLAlchemy crashed (it seems to be the same bug I had when I was working on D972 --- SA doesn't like much extending `__table_args__` in a child class).

Test Plan: Authed and synced accounts. Ran migration.

Reviewers: kav-ya, emfree

Reviewed By: emfree

Projects: #sync-engine

Differential Revision: https://review.inboxapp.com/D1114