Expose 'busy' bit in events API.
Summary:
While we've synced this data forever, we just haven't been serializing it
in API responses. Exposing it is a prerequisite to allowing us to implement
a free/busy API, or for developers to implement their own client-side.

This commit also adds support for setting the 'busy' bit while creating
and updating events, which was supported on the backends but not in the
actual API calls.

You can also now filter events by busy=true/false in the `/events` endpoint.

Test Plan: unit tests updated

Reviewers: khamidou

Reviewed By: khamidou

Subscribers: emfree

Differential Revision: https://review.inboxapp.com/D1250