Add long-polling option for delta sync endpoint
Summary:
Web browsers can't parse partial responses, which makes the streaming endpoint useless to them. This change adds a URL param to enable long-polling. This will effectively "hang" the connection open for 2 minutes or until there is a new change.

Rob Ochshorn wanted something like this. I think it makes a lot of sense to add.

Test Plan: I tested this manually, but I'm not sure how to write a test for it given the async nature.

Reviewers: spang, emfree

Reviewed By: emfree

Subscribers: spang

Differential Revision: https://review.inboxapp.com/D1302