Associate events imported from iCalendar files to the message they were attached to
Summary:
Here's a quick explanation of the why and how. In the previous version of the code we would import iCalendar events in `Message.create_from_synced` (https://github.com/inboxapp/inbox/blob/master/inbox/models/message.py#L272). Unfortunately, at this point in the code no object is flushed to the db yet. This isn't a problem when we're creating events but it is when doing updates. We would have to maintain some per-user event table which would be super complicated and not really worth the effort.

What I chose to do instead was to move the calendar import code a bit lower in the object creation lifecycle --- after flushing a message for the first time. This way we don't have to worry about objects which haven't been flushed yet. The location in `create_message()` was chosen to avoid code duplication with the different IMAP backends.

Test Plan: Unit tests included.

Reviewers: jennie

Reviewed By: jennie

Subscribers: emfree

Projects: #sync-engine

Differential Revision: https://review.inboxapp.com/D1322