Only run "preflight" checks when prod flag is off.
Summary:
The primary motivation for this is to remove a bit of the pain around
concurrently deploying code and schema changes. In nearly all cases, refusing
to start services when there is an alembic version mismatch is
counterproductive to what we actually want to do. E.g., if we're adding a
column we want to first add the column, then update the code. The opposite if
we're dropping a column.

We retain this check as a convenience for developers who may be juggling
branches and schemas, upgrading local VMs, etc.

Test Plan: Run service scripts with and without --prod flag.

Reviewers: jennie

Reviewed By: jennie

Differential Revision: https://review.inboxapp.com/D1343