Fix timestamp-dependent tests.
This fixes MySQL-version-dependent problems with some of the unit tests. These
problems all center around the fact that fractional seconds are discarded from
timestamp fields in MySQL < 5.6.4, but *rounded* in MySQL >= 5.6.4. See
http://dev.mysql.com/doc/refman/5.6/en/fractional-seconds.html. This means that
you can see something like
```
>>> ts = datetime.datetime(2015, 03, 26, 12, 22, 44, 777)
>>> m = Message(...)
>>> m.deleted_at = ts
>>> db_session.commit()
>>> db_session.query(Message).filter(Message.deleted_at <= ts).all()
[]
```

For now we hack around this in tests, but we probably want a better solution in
code (e.g., truncating fractional seconds ourselves prior to insert) to make
this sort of counterintuitive behavior impossible.