[iCal] Merge events participants
Summary:
So, this is a somewhat complicated diff. Here’s a short walkthrough.

The iCalendar protocol is a little annoying. For example, let’s say I create a meeting inviting Alice, Bob and Eve. The invite will contain all three. However, if Alice RSVPs to the event, the invite she will send will only list her as a participant. The organizer is responsible for making sense of everything.

There’s a second problem. Let’s say I only know Bob’s address and not his name — some servers, like Exchange, will helpfully generate an iCalendar file with his email address as his name. In this case, we need an algorithm to merge correctly names.

The third problem is that some people may not have an email address, or the organizer doesn’t know it.

The two parts I’d like you to check out first are:
1. the merging algorithm itself. The idea behind it is that:
   - email addresses are mostly unique
   - names are not
   When looking up participants for merging we first look if we have an already existing entry with this email address. If yes, we update it. If no, we use the name field.

A thing I’m unsure about is whether it’s okay to consider that two participants with the same names are the same person. On the one hand, it allows us to fill out a participant’s email address after the fact. On the other hand, if we get two Bob Smiths attending the same meeting, they’ll get merged together. The latter is somewhat mitigated by the fact that since we use RSVP responses to merge participants, we’re almost certain that we have valid email addresses, so this shouldn’t happen.

2. The trick I’m using to merge participants from older RSVPs. It should work but you never know

Sorry for the wall of text.

Test Plan: Tested manually + unit tests.

Reviewers: emfree, spang, jennie

Reviewed By: jennie

Differential Revision: https://review.inboxapp.com/D1381